requirements.txt
# Web3 and crypto dependencies
web3>=6.0.0
python-binance==1.0.19
eth-account==0.8.0

# Utilities and helpers
python-dotenv>=0.19.0
requests==2.32.3
pandas==2.2.2
numpy==1.26.4

# Async support
aiohttp==3.11.10
asyncio==3.4.3

# Telegram bot
python-telegram-bot>=20.0

# Testing
pytest>=8.0.0
